+++ template/connectorType/utilities/GetEntityList.php
@@ -9,6 +9,11 @@
  */
 
 ?>
+<a class='btn btn-link' href='Connecteur/editionModif?id_ce=<?php hecho((string)$id_ce); ?>'>
+    <i class="fa fa-arrow-left"></i>&nbsp;Retour au connecteur
+</a>
+<div class="box">
+    <h2 id="desc-entite-table">Choisissez une entité</h2>
 
 <form action='Connecteur/doExternalData' method='post'>
     <input type='hidden' name='id_ce' value='<?php hecho((string)$id_ce); ?>'/>
@@ -36,3 +41,4 @@
         <i class="fa fa-check"></i>&nbsp;Choisir
     </button>
 </form>
+</div>